Move upload logic from storage folder to internal.
I want to classify classes according to the reMarkable API endpoints. The file upload logic does not rely on the documentStorage endpoint, but in the internal.
