+++ src/internal/FileBuffer.ts
