+++ src/internal/FileBufferType.ts
